Try to find lock(s) controlling scheduler
This patch includes my changes in attempt to find the locks that control
oneTBB's scheduler and the program I used to test those changes, under
the demo directory. It is intended to be applied over commit 8155aae.

My changes include replacing several mutexes with null_mutexes. This
should effectively removes those locks. Specifially, I replaced
__TBB_SCHEDULER_MUTEX_TYPE defined in scheduler_common.h (and thus the
other types defined in terms of it), the queue mutex in task_stream.h,
and the task group context list mutex in thread_data.h. Even if all of
these mutexes are not directly related to scheduler, I expected to see
some difference in execution with these changes. To make sure the linker
was using this modified version of TBB, I inserted print statements around
the source files I touched. I built TBB using the following commands:

	cmake .
	cmake --build . -j$(nproc)

I tested these changes with at test program that recursively computes
Fibonacci numbers via parallel_invoke. I put it, other source files it
relies on, and a makefile in the demo directory. It can be built like so:

	cd demo
	make

I ran the program using the following commands (in the demo directory):

	echo normal
	echo 42 $(nproc) | ./recursive-fib 5
	echo modified
	echo 42 $(nproc) | LD_LIBRARY_PATH=../gnu_10.3_cxx1_64_relwithdebinfo ./recursive-fib 5

The process running with the modified TBB would output the print
statements I inserted as expected. However, after diverting standard error
to /dev/null, the output was the same (i.e., the Fibonacci computations
were correct) as the one running with the normal, system-wide version
of TBB.

All code was compiled and tested on the eightsocket machine.